Server: send 'isSupportSession' flag as window property instead of Redux state
On client, we need to read the value of the `isSupportSession` flag before initialization
of the Redux store. It decides whether the persisted state is or is not read from local
storage.

It's very similar to how `window.currentUser` is sent. It's also needed before Redux store
initialization (persisted state has the user ID as part of the key) and we therefore don't
send it as part of initial state, but as a separate variable.